Refactors settings page and moves buttons
