+++ src/css/DegreeSettingsPage.css
@@ -4,14 +4,25 @@
   margin-top: 36px;
 }
 
+.page-body {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
 .page-actions {
+  display: flex;
+  justify-content: flex-end;
   width: 100%;
-  margin-bottom: 64px;
 }
 
 .page-sidebar {
   height: 80vh;
   border: 1px solid black;
+}
+
+.degree-info {
+  margin: 4rem 0rem;
 }
 
 .degree-info > div:not(:first-child) {
